[TASK] Refactor console setup on composer install
Since we're going to do more stuff on build time,
we need an improved API and get rid of the huge class
